+++ pkg/nexus/search.go
@@ -2,6 +2,7 @@ package nexus
 
 import (
 	"fmt"
+	"strings"
 
 	nexusrm "github.com/sonatype-nexus-community/gonexus/rm"
 )
@@ -17,7 +18,14 @@ func (c *Client) Search(repository, group string) ([]string, error) {
 
 	res := []string{}
 	for _, a := range assets {
-		res = append(res, a.DownloadURL)
+		res = append(res, appendTestTLSPort(a.DownloadURL))
 	}
 	return res, nil
+}
+
+// appendTestTLSPort appends the TLS port used in testing to the URL.
+func appendTestTLSPort(in string) string {
+	return strings.Replace(
+		strings.Replace(in, "localhost/", "localhost:8443/", -1),
+		"ods-test-nexus-tls.kind/", "ods-test-nexus-tls.kind:8443/", -1)
 }